backend/requirements.txt
fastapi
uvicorn
prisma
fastapi-auth0
python-multipart
google-cloud-storage
